pyproject.toml
[tool.poetry]
name = "pomodoro_bot"
version = "0.1.0"
description = "Telegram bot for Pomodoro technique"
authors = ["Developer <developer@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = {version = "^20.0", extras = ["webhooks"]}
SQLAlchemy = "^2.0"
python-dotenv = "^1.0"
asyncpg = "^0.29.0"
apscheduler = "^3.10.4"
pytz = "^2024.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.8"
black = "^23.12.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "I"]
line-length = 88
target-version = "py310"
exclude = [".git", ".venv", "__pycache__", ".pytest_cache"]

[tool.black]
line-length = 88
target-version = ["py310"]
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | \.pytest_cache
)/
''' 